useNewsroomContext and NewsroomContextProvider
